fixed wand length data type from int to double

+++ app/src/main/java/com/saadahmedev/hpcapi/data/dto/Wand.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class Wand(
     @SerializedName("wood") var wood: String? = null,
     @SerializedName("core") var core: String? = null,
-    @SerializedName("length") var length: Int? = null
+    @SerializedName("length") var length: Double? = null
 )
